requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==1.10.13
python-jose[cryptography]==3.3.0
motor==3.4.0
dnspython==2.4.2
loguru==0.7.2
bcrypt==3.2.2
pydantic[email]<2
python-multipart==0.0.9
